Update browserAction.rst
make users aware of CSP. There are so many posts asking to use html but not pointing to CSP. So the numerous MDN examples (e.g. for File input) don't work in popup.html/js context because of CSP. I wasted a full evening on this until someone pointed me there.

It is not obvious that html does not work in popup when I am asked to replace xul by html.